% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGSscraper.R
\name{rungreylitsearcher}
\alias{rungreylitsearcher}
\title{Wrapper to run greylitsearcher Shiny app locally}
\usage{
rungreylitsearcher()
}
\value{
Runs the GitHub version of the Shiny App locally.
}
\description{
Run the greylitsearcher Shiny app locally with this
function. Calls the latest version of the app from GitHub without
the need to install greylitsearcher
}
\examples{
\dontrun{
rungreylitsearcher()
}
}
